# openapi/openapi.json
{
  "components" : {
    "schemas" : {
...
      "ApiResponseCommentCreated" : {
        "properties" : {
          "data" : {
            "$ref" : "#/components/schemas/CommentCreated"
          },
          "error" : {
            "$ref" : "#/components/schemas/ErrorResponse"
          },
          "success" : {
            "type" : "boolean"
          },
          "timestamp" : {
            "format" : "date-time",
            "type" : "string"
          },
          "traceId" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "ApiResponseConsultingLeadResponse" : {
        "properties" : {
          "data" : {
            "$ref" : "#/components/schemas/ConsultingLeadResponse"
          },
          "error" : {
            "$ref" : "#/components/schemas/ErrorResponse"
          },
          "success" : {
            "type" : "boolean"
          },
          "timestamp" : {
            "format" : "date-time",
            "type" : "string"
          },
          "traceId" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "ApiResponsePostCreated" : {
        "properties" : {
          "data" : {
            "$ref" : "#/components/schemas/PostCreated"
          },
          "error" : {
            "$ref" : "#/components/schemas/ErrorResponse"
          },
          "success" : {
            "type" : "boolean"
          },
          "timestamp" : {
            "format" : "date-time",
            "type" : "string"
          },
          "traceId" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "ApiResponsePostList" : {
        "properties" : {
          "data" : {
            "$ref" : "#/components/schemas/PostList"
          },
          "error" : {
            "$ref" : "#/components/schemas/ErrorResponse"
          },
          "success" : {
            "type" : "boolean"
          },
          "timestamp" : {
            "format" : "date-time",
            "type" : "string"
          },
          "traceId" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "ApiResponsePostView" : {
        "properties" : {
          "data" : {
            "$ref" : "#/components/schemas/PostView"
          },
          "error" : {
            "$ref" : "#/components/schemas/ErrorResponse"
          },
          "success" : {
            "type" : "boolean"
          },
          "timestamp" : {
            "format" : "date-time",
            "type" : "string"
          },
          "traceId" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "ApiResponseVoid" : {
        "properties" : {
          "data" : {
            "type" : "object"
          },
          "error" : {
            "$ref" : "#/components/schemas/ErrorResponse"
          },
          "success" : {
            "type" : "boolean"
          },
          "timestamp" : {
            "format" : "date-time",
            "type" : "string"
          },
          "traceId" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "AttachmentView" : {
        "properties" : {
          "contentType" : {
            "type" : "string"
          },
          "downloadUrl" : {
            "type" : "string"
          },
          "fileId" : {
            "type" : "string"
          },
          "originalName" : {
            "type" : "string"
          },
          "sizeInBytes" : {
            "format" : "int64",
            "type" : "integer"
          }
        },
        "type" : "object"
      },
...
      "CommentCreated" : {
        "properties" : {
          "commentId" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "CommentItem" : {
        "properties" : {
          "authorNickname" : {
            "type" : "string"
          },
          "body" : {
            "type" : "string"
          },
          "commentId" : {
            "type" : "string"
          },
          "createdAt" : {
            "type" : "string"
          },
          "editable" : {
            "type" : "boolean"
          }
        },
        "type" : "object"
      },
...
      "ConsultingLeadResponse" : {
        "properties" : {
          "diagnosisId" : {
            "type" : "string"
          },
          "id" : {
            "type" : "string"
          },
          "maskedEmail" : {
            "type" : "string"
          },
          "maskedPhone" : {
            "type" : "string"
          },
          "status" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "ErrorResponse" : {
        "properties" : {
          "code" : {
            "type" : "string"
          },
          "details" : {
            "additionalProperties" : {
              "type" : "object"
            },
            "type" : "object"
          },
          "fieldErrors" : {
            "items" : {
              "$ref" : "#/components/schemas/FieldError"
            },
            "type" : "array"
          },
          "message" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "FieldError" : {
        "properties" : {
          "field" : {
            "type" : "string"
          },
          "reason" : {
            "type" : "string"
          },
          "rejectedValue" : {
            "type" : "object"
          }
        },
        "type" : "object"
      },
...
      "PostCreated" : {
        "properties" : {
          "postId" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "PostList" : {
        "properties" : {
          "page" : {
            "$ref" : "#/components/schemas/PostPage"
          }
        },
        "type" : "object"
      },
      "PostPage" : {
        "properties" : {
          "page" : {
            "format" : "int32",
            "type" : "integer"
          },
          "posts" : {
            "items" : {
              "$ref" : "#/components/schemas/PostSummary"
            },
            "type" : "array"
          },
          "size" : {
            "format" : "int32",
            "type" : "integer"
          },
          "totalCount" : {
            "format" : "int64",
            "type" : "integer"
          }
        },
        "type" : "object"
      },
      "PostSummary" : {
        "properties" : {
          "authorNickname" : {
            "type" : "string"
          },
          "commentCount" : {
            "format" : "int32",
            "type" : "integer"
          },
          "createdAt" : {
            "type" : "string"
          },
          "maskedTitle" : {
            "type" : "string"
          },
          "postId" : {
            "type" : "string"
          },
          "readable" : {
            "type" : "boolean"
          },
          "secret" : {
            "type" : "boolean"
          },
          "viewCount" : {
            "format" : "int64",
            "type" : "integer"
          }
        },
        "type" : "object"
      },
      "PostView" : {
        "properties" : {
          "attachments" : {
            "items" : {
              "$ref" : "#/components/schemas/AttachmentView"
            },
            "type" : "array"
          },
          "authorNickname" : {
            "type" : "string"
          },
          "boardType" : {
            "type" : "string"
          },
          "body" : {
            "type" : "string"
          },
          "comments" : {
            "items" : {
              "$ref" : "#/components/schemas/CommentItem"
            },
            "type" : "array"
          },
          "createdAt" : {
            "type" : "string"
          },
          "editable" : {
            "type" : "boolean"
          },
          "postId" : {
            "type" : "string"
          },
          "secret" : {
            "type" : "boolean"
          },
          "title" : {
            "type" : "string"
          },
          "updatedAt" : {
            "type" : "string"
          },
          "viewCount" : {
            "format" : "int64",
            "type" : "integer"
          }
        },
        "type" : "object"
      },
...
      "RequestConsultingRequest" : {
        "properties" : {
          "consentVersion" : {
            "minLength" : 1,
            "type" : "string"
          },
          "contactEmail" : {
            "type" : "string"
          },
          "contactName" : {
            "minLength" : 1,
            "type" : "string"
          },
          "contactPhone" : {
            "minLength" : 1,
            "type" : "string"
          },
          "diagnosisId" : {
            "format" : "int64",
            "type" : "integer"
          },
          "message" : {
            "type" : "string"
          },
          "privacyConsent" : {
            "type" : "boolean"
          },
          "sensitiveInfoConsent" : {
            "type" : "boolean"
          },
          "serviceLimitAcknowledged" : {
            "type" : "boolean"
          }
        },
        "required" : [ "diagnosisId" ],
        "type" : "object"
      },
...
      "WriteComment" : {
        "properties" : {
          "body" : {
            "maxLength" : 2000,
            "minLength" : 0,
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "WritePost" : {
        "properties" : {
          "attachmentFileIds" : {
            "items" : {
              "type" : "string"
            },
            "type" : "array"
          },
          "body" : {
            "minLength" : 1,
            "type" : "string"
          },
          "secret" : {
            "type" : "boolean"
          },
          "title" : {
            "maxLength" : 200,
            "minLength" : 0,
            "type" : "string"
          }
        },
        "type" : "object"
      }
    },
    "securitySchemes" : {
      "bearerAuth" : {
        "bearerFormat" : "JWT",
        "description" : "로그인(POST /api/v1/auth/login) 응답의 accessToken",
        "scheme" : "bearer",
        "type" : "http"
      }
    }
  },
  "info" : {
    "description" : "1인 제조기업·소공인 제품 인증 준비도 자가진단 및 컨설팅 연계 서비스.\n\n• 모든 응답은 ApiResponse 봉투(success/data/error/traceId/timestamp)로 감쌉니다.\n• 준비도 점수는 인증 합격·불합격 판정이 아니라 준비 상태를 확인하는 사전 점검 지표입니다.\n• 인증이 필요한 엔드포인트는 우상단 Authorize에 로그인 accessToken을 넣고 호출하세요.\n",
    "license" : {
      "name" : "팀 육성사이다"
    },
    "title" : "OpenCertFlow API",
    "version" : "v1"
  },
  "openapi" : "3.1.0",
  "paths" : {
...
    "/api/v1/boards/comments/{commentId}" : {
      "delete" : {
        "operationId" : "deleteComment",
        "parameters" : [ {
          "in" : "path",
          "name" : "commentId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseVoid"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "board-controller" ]
      },
      "put" : {
        "operationId" : "editComment",
        "parameters" : [ {
          "in" : "path",
          "name" : "commentId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/WriteComment"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseVoid"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "board-controller" ]
      }
    },
    "/api/v1/boards/posts/{postId}" : {
      "delete" : {
        "operationId" : "delete_1",
        "parameters" : [ {
          "in" : "path",
          "name" : "postId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseVoid"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "board-controller" ]
      },
      "get" : {
        "operationId" : "get_1",
        "parameters" : [ {
          "in" : "path",
          "name" : "postId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponsePostView"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "board-controller" ]
      },
      "put" : {
        "operationId" : "edit",
        "parameters" : [ {
          "in" : "path",
          "name" : "postId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/WritePost"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponsePostCreated"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "board-controller" ]
      }
    },
    "/api/v1/boards/posts/{postId}/comments" : {
      "post" : {
        "operationId" : "writeComment",
        "parameters" : [ {
          "in" : "path",
          "name" : "postId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/WriteComment"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseCommentCreated"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "board-controller" ]
      }
    },
...
    "/api/v1/boards/{boardType}/posts" : {
      "get" : {
        "operationId" : "list",
        "parameters" : [ {
          "in" : "path",
          "name" : "boardType",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "page",
          "required" : false,
          "schema" : {
            "default" : 0,
            "format" : "int32",
            "type" : "integer"
          }
        }, {
          "in" : "query",
          "name" : "size",
          "required" : false,
          "schema" : {
            "default" : 20,
            "format" : "int32",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponsePostList"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "board-controller" ]
      },
      "post" : {
        "operationId" : "write",
        "parameters" : [ {
          "in" : "path",
          "name" : "boardType",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/WritePost"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponsePostCreated"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "board-controller" ]
      }
    },
    "/api/v1/consulting-leads" : {
      "post" : {
        "operationId" : "request",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/RequestConsultingRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseConsultingLeadResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "상담 신청(공개)" ]
      }
    },
...
  },
  "security" : [ {
    "bearerAuth" : [ ]
  } ],
  "servers" : [ {
    "description" : "Generated server url",
    "url" : ""
  } ],
...
}